requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html 
torch==1.10.2+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.10.2+cu113
pybind11
pkbar==0.5
fastwer==0.1.3
audiofile==1.0.3
torch-summary==1.4.5
Execution-Time==1.1.0
tqdm==4.62.3
twine==3.7.1
Cython==0.29.26
scipy==1.7.3
sentencepiece==0.1.96
flashtext==2.7
numpy==1.21.4
numba==0.53
pydub==0.25.1
torch-summary==1.4.5
torch-audiomentations==0.9.1
pyctcdecode==0.1.1
alive-progress==2.0.0
more_itertools==8.10.0
prettytable==2.2.1
pyyaml==5.4.1
tensorboard==2.6.0
prettytable==2.2.1
unidecode==1.3.4
asr-evaluation
omegaconf
hydra-core==1.2.0
bitarray==2.5.1
sacrebleu==2.1.0
tensorboardX==2.5.1
npy_append_array
p_tqdm
torchmetrics==0.9.2
fast-pytorch-kmeans
git+https://github.com/kpu/kenlm@master